Add "sample" root package (for "proto-lookup" plugin).
