+++ common/src/main/proto/lobby/contracts/common/common.proto
@@ -21,7 +21,7 @@ syntax = "proto3";
 
 package lobby.contracts.common;
 
-option java_package="lobby.contracts.common";
+option java_package="sample.lobby.contracts.common";
 option java_outer_classname = "CommonContractsProto";
 option java_multiple_files = true;
