+++ conference-contracts/src/main/proto/lobby/contracts/conference/events.proto
@@ -21,7 +21,7 @@ syntax = "proto3";
 
 package lobby.contracts.conference;
 
-option java_package="lobby.contracts.conference";
+option java_package="sample.lobby.contracts.conference";
 option java_outer_classname = "ConferenceContractsEventsProto";
 option java_multiple_files = true;
