+++ payment-contracts/src/main/proto/lobby/contracts/payment/events.proto
@@ -21,7 +21,7 @@ syntax = "proto3";
 
 package lobby.contracts.payment;
 
-option java_package="lobby.contracts.payment";
+option java_package="sample.lobby.contracts.payment";
 option java_outer_classname = "PaymentContractsEventsProto";
 option java_multiple_files = true;
