+++ registration-contracts/src/main/proto/lobby/contracts/registration/events.proto
@@ -21,7 +21,7 @@ syntax = "proto3";
 
 package lobby.contracts.registration.order;
 
-option java_package="lobby.contracts.registration.order";
+option java_package="sample.lobby.contracts.registration.order";
 option java_outer_classname = "RegistrationContractsEventsProto";
 option java_multiple_files = true;
