+++ registration-contracts/src/main/proto/lobby/contracts/registration/registration.proto
@@ -21,7 +21,7 @@ syntax = "proto3";
 
 package lobby.contracts.registration;
 
-option java_package="lobby.contracts.registration";
+option java_package="sample.lobby.contracts.registration";
 option java_outer_classname = "RegistrationContractsProto";
 option java_multiple_files = true;
