+++ registration/src/main/java/sample/lobby/registration/package-info.java
@@ -19,6 +19,6 @@
  */
 
 @ParametersAreNonnullByDefault
-package lobby.registration;
+package sample.lobby.registration;
 
 import javax.annotation.ParametersAreNonnullByDefault;
