+++ registration/src/main/java/sample/lobby/registration/service/package-info.java
@@ -19,6 +19,6 @@
  */
 
 @ParametersAreNonnullByDefault
-package lobby.registration.projection;
+package sample.lobby.registration.service;
 
 import javax.annotation.ParametersAreNonnullByDefault;
