+++ registration/src/main/proto/lobby/registration/conference.proto
@@ -21,7 +21,7 @@ syntax = "proto3";
 
 package lobby.registration;
 
-option java_package="lobby.registration";
+option java_package="sample.lobby.registration";
 option java_outer_classname = "ConferenceRegistrationProto";
 option java_multiple_files = true;
