+++ registration/src/main/proto/lobby/registration/order/order.proto
@@ -21,7 +21,7 @@ syntax = "proto3";
 
 package lobby.registration.order;
 
-option java_package="lobby.registration.order";
+option java_package="sample.lobby.registration.order";
 option java_outer_classname = "OrderProto";
 option java_multiple_files = true;
