+++ registration/src/main/proto/lobby/registration/process/process.proto
@@ -21,7 +21,7 @@ syntax = "proto3";
 
 package lobby.registration.process;
 
-option java_package="lobby.registration.process";
+option java_package="sample.lobby.registration.process";
 option java_outer_classname = "RegistrationProcessProto";
 option java_multiple_files = true;
